requirements.txt
numpy==1.26.4
pandas==2.1.4
scipy==1.11.4
scikit-learn==1.5.1
xgboost==2.0.3
tensorflow-macos==2.16.2
tensorflow-metal==1.2.0
matplotlib==3.8.2
plotly==5.18.0
seaborn==0.13.1
streamlit==1.31.0
pyyaml==6.0.1
joblib==1.3.2
tqdm==4.66.1
